Improve tests and add vcr and webmock

+++ test/helper.rb
@@ -5,6 +5,9 @@ require 'bigquery-client'
 require 'dotenv'
 require 'awesome_print'
 require 'pry-byebug'
+require 'vcr'
+
+Dotenv.load
 
 if defined?(PryByebug)
   Pry.commands.alias_command 'c', 'continue'
@@ -13,9 +16,13 @@ if defined?(PryByebug)
   Pry.commands.alias_command 'f', 'finish'
 end
 
-Dotenv.load
+VCR.configure do |config|
+  config.cassette_library_dir = "vcr_cassettes"
+  config.hook_into :webmock
+end
 
-$dataset = 'test_bigquery_client'
+$prefix  = 'test_bigquery_client'
+$dataset = "#{$prefix}_default"
 
 $client = BigQuery::Client.new(
   project:                ENV['BIGQUERY_PROJECT'],
@@ -26,10 +33,19 @@ $client = BigQuery::Client.new(
   auth_method:            'private_key'
 )
 
-$client.datasets.select { |dataset|
-  Regexp.new($dataset) =~ dataset
-}.each { |dataset|
-  $client.delete_dataset(dataset)
-}
+VCR.use_cassette("helper") do
+  $client.datasets.each do |dataset|
+    $client.delete_dataset(dataset) if Regexp.new($prefix) =~ dataset
+  end
+  $client.create_dataset($dataset)
+end
 
-$client.create_dataset($dataset)
+Test::Unit::TestCase.class_eval do
+  def setup
+    VCR.insert_cassette("#{self.class}_#{@method_name}")
+  end
+
+  def cleanup
+    VCR.eject_cassette
+  end
+end
